1.02b

+++ experimental/distributed_fuzzing/sync_script.sh
@@ -41,6 +41,11 @@ SYNC_DIR='/home/bob/sync_dir'
 # Interval (seconds) between sync attempts
 SYNC_INTERVAL=$((30 * 60))
 
+if [ "$PWD" = "/tmp" -o "$PWD" = "/var/tmp" ]; then
+  echo "Error: do not use shared /tmp or /var/tmp directories with this script." 1>&2
+  exit 1
+fi
+
 rm -rf .sync_tmp 2>/dev/null
 mkdir .sync_tmp || exit 1
